backend/requirements.txt
fastapi
uvicorn[standard]
dask[complete]
pandas
scikit-learn
python-dateutil
joblib
kaggle